+++ openssl/src/crypto/asn1/a_time.c
@@ -16,7 +16,7 @@
 
 #include <stdio.h>
 #include <time.h>
-#include "internal/ctype.h"
+#include "internal/ossl_type.h"
 #include "internal/cryptlib.h"
 #include <openssl/asn1t.h>
 #include "asn1_locl.h"
