+++ openssl/src/crypto/ctype.c
@@ -9,7 +9,7 @@
 
 #include <string.h>
 #include <stdio.h>
-#include "internal/ctype.h"
+#include "internal/ossl_type.h"
 #include "openssl/ebcdic.h"
 
 /*
