+++ openssl/src/crypto/pem/pem_lib.c
@@ -8,7 +8,7 @@
  */
 
 #include <stdio.h>
-#include "internal/ctype.h"
+#include "internal/ossl_type.h"
 #include <string.h>
 #include "internal/cryptlib.h"
 #include <openssl/buffer.h>
