+++ openssl/src/crypto/store/loader_file.c
@@ -25,7 +25,7 @@
 #include <openssl/ui.h>
 #include <openssl/x509.h>        /* For the PKCS8 stuff o.O */
 #include "internal/asn1_int.h"
-#include "internal/ctype.h"
+#include "internal/ossl_type.h"
 #include "internal/o_dir.h"
 #include "internal/cryptlib.h"
 #include "internal/store_int.h"
